+++ services/cache/cache.go
@@ -2,5 +2,5 @@ package cache
 
 // Cache defines interface that one should implement
 type Cache interface {
-	GetBitcoinPrice() float64
+	GetBitcoinPrice() int64
 }
